refactor(menubar): split menu bar view sections
