+++ ArcBox/Views/MenuBar/MenuBarView/MenuBarHoverButton.swift
@@ -1,0 +1,25 @@
+import SwiftUI
+
+// MARK: - Hover Components
+
+struct MenuBarHoverButton<Label: View>: View {
+    var cornerRadius: CGFloat = 6
+    let action: () -> Void
+    @ViewBuilder let label: Label
+
+    @State private var isHovering = false
+
+    var body: some View {
+        Button(action: action) {
+            label
+                .frame(maxWidth: .infinity, alignment: .leading)
+                .background(
+                    RoundedRectangle(cornerRadius: cornerRadius, style: .continuous)
+                        .fill(isHovering ? Color.primary.opacity(0.10) : .clear)
+                )
+                .contentShape(Rectangle())
+        }
+        .buttonStyle(.plain)
+        .onHover { isHovering = $0 }
+    }
+}
